api/src/requirements.txt
fastapi==0.101.1
ruamel.yaml==0.17.32
uvicorn==0.23.2
PyYAML==6.0.1
